use programs offered rather than attended per #566
This commit introduces a new field, `picc.fields.PROGRAM_OFFERED` and
changes the behavior of `picc.access.programAreas()` so that it can take
an optional second argument of the field to use for accessing the
program listing. The default is `picc.fields.PROGRAM_OFFERED`, but for
the `popular_programs` directive we use
`picc.fields.PROGRAM_PERCENTAGE`. The `available_programs` directive is
unchanged, but now lists all of the programs offered rather than just
the ones attended in 2013.

We also now use the `picc.PROGRAM_AREAS_BY_ID` mapping set in
`_includes/js-config.html` (included in the `<head>` of every page) to
determine human-friendly labels for the API fields, rather than the API
metadata dictionary, since those labels don't exist for both fields.